Add files via upload
Particle Swarm Optimization (PSO) is a computational optimization technique inspired by the social behaviour of birds or fish flocking. It's used to find optimal solutions to complex problems, particularly in fields such as engineering, computer science, and mathematics. In PSO, a group of potential solutions, called particles, move through a solution space to search for the best solution. Each particle represents a potential solution and has a position and velocity. The particles' movements are guided by their own experience and the experience of their neighbours in the search space.